Fixed use of the evil long type in Q_rsqrt causing a lots of glitches in renderer, changed some other longs to ints as well.

+++ source/botlib/botlib.h
@@ -196,7 +196,7 @@ typedef struct botlib_import_s
 	int (* FS_Read)(void* buffer, int len, fileHandle_t f);
 	int (* FS_Write)(const void* buffer, int len, fileHandle_t f);
 	void (* FS_FCloseFile)(fileHandle_t f);
-	int (* FS_Seek)(fileHandle_t f, long offset, int origin);
+	int (* FS_Seek)(fileHandle_t f, int offset, int origin);
 	//debug visualisation stuff
 	int (* DebugLineCreate)(void);
 	void (* DebugLineDelete)(int line);
