+++ source/common/mathlib.cpp
@@ -191,13 +191,13 @@ float bytedirs[NUMVERTEXNORMALS][3] =
 
 float Q_rsqrt(float number)
 {
-	long i;
+	int i;
 	float x2, y;
 	const float threehalfs = 1.5F;
 
 	x2 = number * 0.5F;
 	y  = number;
-	i  = *(long*)&y;						// evil floating point bit level hacking
+	i  = *(int*)&y;						// evil floating point bit level hacking
 	i  = 0x5f3759df - (i >> 1);				// what the fuck?
 	y  = *(float*)&i;
 	y  = y * (threehalfs - (x2 * y * y));	// 1st iteration
@@ -252,7 +252,7 @@ float Q_acos(float c)
 
 #if id386 && defined _MSC_VER
 #pragma warning (disable:4035)
-__declspec(naked) long Q_ftol(float f)
+__declspec(naked) int Q_ftol(float f)
 {
 	static int tmp;
 	__asm fld dword ptr [esp + 4]
